move to convention 1.0.0
