+++ .all-contributorsrc
@@ -1,0 +1,16 @@
+{
+  "projectName": "module_euro_gas_grid",
+  "projectOwner": "modelblocks-org",
+  "repoType": "github",
+  "repoHost": "https://github.com",
+  "files": [
+    "README.md"
+  ],
+  "imageSize": 100,
+  "commit": false,
+  "commitConvention": "none",
+  "contributors": [],
+  "contributorsPerLine": 7,
+  "linkToUsage": false,
+  "contributorsSortAlphabetically": true
+}
