+++ .copier-answers.yml
@@ -1,12 +1,13 @@
 # Changes here will be overwritten by Copier
 # !!!!! DO NOT MANUALLY MODIFY THIS FILE !!!!!
-_commit: latest
+_commit: v1.0.2
 _src_path: https://github.com/modelblocks-org/data-module-template.git
 author_email: i.ruizmanuel@tudelft.nl
 author_family_name: Ruiz Manuel
 author_given_name: Ivan
 github_org: modelblocks-org
 license: Apache-2.0
-module_description: A module to cluster European gas networks into any resolution.
-module_long_name: Module Euro Gas Grid
+module_description: A module to cluster European gas networks from SciGRID_gas into
+    any resolution.
+module_long_name: Euro Gas Grid Module
 module_short_name: module_euro_gas_grid
