+++ .github/workflows/pr-ci.yml
@@ -13,11 +13,10 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
-        python-version: ["3.12"]
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v7
       - name: Setup pixi
-        uses: prefix-dev/setup-pixi@v0.8.3
+        uses: prefix-dev/setup-pixi@v0.9.6
       - name: Run integration tests
         id: tests
         run: pixi run test-integration
@@ -30,10 +29,10 @@ jobs:
         continue-on-error: true
       - name: Save integration logs
         if: ${{ always() }}
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v7
         with:
           name: integration-test-logs-${{ matrix.os }}
-          path: tests/integration/results/integration_test/logs
+          path: tests/integration/resources/module/logs
           if-no-files-found: ignore
           retention-days: 30
       - name: Fail if integration or linting failed
