--- .github/workflows/release.yml
@@ -1,11 +1,0 @@
-name: Release
-
-on:
-  release:
-    types: [released]
-
-jobs:
-  release-workflow:
-    permissions:
-      contents: write
-    uses: modelblocks-org/data-module-template/.github/workflows/template-release.yml@latest
