+++ .pixi/config.toml
@@ -1,0 +1,1 @@
+pinning-strategy = "latest-up"
