+++ CITATION.cff
@@ -3,7 +3,7 @@
 # https://citation-file-format.github.io/
 cff-version: 1.2.0
 message: If you use this software or data produced by it, please cite it using the metadata from this file.
-title: "Modelblocks - module_euro_gas_grid: Module Euro Gas Grid"
+title: "Modelblocks - module_euro_gas_grid: Euro Gas Grid Module"
 repository: "https://github.com/modelblocks-org/module_euro_gas_grid"
 license: Apache-2.0
 authors:
