+++ pixi.toml
@@ -1,7 +1,7 @@
 [workspace]
 name = "module_euro_gas_grid"
 authors = ["See AUTHORS file"]
-description = "A module to cluster European gas networks into any resolution."
+description = "A module to cluster European gas networks from SciGRID_gas into any resolution."
 license = "Apache-2.0"
 readme = "README.md"
 channels = ["conda-forge", "bioconda"]
@@ -9,7 +9,7 @@ platforms = ["win-64", "linux-64", "osx-arm64"]
 homepage = "https://www.modelblocks.org/"
 
 [dependencies]
-clio-tools = ">=2026.03.30"
+clio-tools = ">=2026.06.23"
 conda = ">=25.0.0"
 ipdb = ">=0.13.13"
 ipykernel = ">=6.29.5"
@@ -19,18 +19,13 @@ pytest = ">=8.3.5"
 python = ">=3.12"
 ruff = ">=0.9.9"
 snakefmt = ">=0.10.2"
-snakemake-minimal = ">=8.29.0"
+snakemake-minimal = ">=9.19.0"
 pytz = ">=2026.1.post1"
 
-[tasks]
-test-integration = {cmd = "pytest tests/integration_test.py"}
-
-[feature.euro-gas-grid]
-channels = ["conda-forge", "nodefaults"]
-
-[feature.euro-gas-grid.dependencies]
+[feature.module.dependencies]
 cmap = "0.6.2.*"
 country_converter = "1.3.2.*"
+curl = ">=8.9.1"
 geopandas = "1.0.1.*"
 matplotlib = "3.10.8.*"
 numpy = "2.3.4.*"
@@ -43,4 +38,28 @@ scipy = "1.16.3.*"
 shapely = "2.1.2.*"
 
 [environments]
-euro-gas-grid = { features = ["euro-gas-grid"], no-default-feature = true }
+module = { features = ["module"], no-default-feature = true }
+
+[tasks]
+test-integration = {cmd = "pytest tests/integration_test.py"}
+
+
+[tasks.export-snakemake-env]
+description = "Export one Pixi environment as Snakemake-compatible conda files"
+args = ["env", { arg = "outdir", default = "workflow/envs" }]
+depends-on = [
+  { task = "_export-snakemake-pin", args = ["{{ env }}", "win-64", "{{ outdir }}"] },
+  { task = "_export-snakemake-pin", args = ["{{ env }}", "linux-64", "{{ outdir }}"] },
+  { task = "_export-snakemake-pin", args = ["{{ env }}", "osx-arm64", "{{ outdir }}"] },
+]
+cmd = "pixi workspace export conda-environment --environment '{{ env }}' '{{ outdir }}/{{ env }}.yaml'"
+
+[tasks._export-snakemake-pin]
+description = "Export one Pixi environment/platform as a Snakemake-compatible pin file"
+args = ["env", "platform", { arg = "outdir", default = "workflow/envs" }]
+cmd = """
+set -e
+pixi workspace export conda-explicit-spec --environment '{{ env }}' --platform '{{ platform }}' '{{ outdir }}'
+mv '{{ outdir }}/{{ env }}_{{ platform }}_conda_spec.txt' '{{ outdir }}/{{ env }}.{{ platform }}.pin.txt'
+"""
+
