+++ tests/conftest.py
@@ -10,6 +10,12 @@ import pytest
 TEST_FILES = "https://zenodo.org/records/18153129/files/test_suite.zip?download=1"
 
 
+@pytest.fixture(scope="module")
+def module_path():
+    """Parent directory of the project."""
+    return Path(__file__).parent.parent
+
+
 @pytest.fixture(scope="session")
 def user_path() -> Path:
     """Download and unzip test files."""
@@ -24,3 +30,20 @@ def user_path() -> Path:
         with zipfile.ZipFile(test_zip, "r") as zfile:
             zfile.extractall(dir)
     return dir / "user"
+
+
+@pytest.fixture(scope="module")
+def integration_path(user_path: Path, module_path: Path):
+    """Ensures the minimal integration test is ready."""
+    integration_dir = Path(module_path / "tests/integration")
+    if integration_dir.exists():
+        shutil.rmtree(
+            integration_dir / "results/", ignore_errors=True
+        )  # clean everything
+    user_integ_dir = integration_dir / "resources/inputs/"
+    files_to_copy = ["BALK/shapes.parquet"]
+    for file in files_to_copy:
+        destination_file = Path(user_integ_dir / file)
+        destination_file.parent.mkdir(parents=True, exist_ok=True)
+        shutil.copy(user_path / file, destination_file)
+    return integration_dir
