--- workflow/envs/euro-gas-grid.yaml
@@ -1,17 +1,0 @@
-name: euro-gas-grid
-channels:
-  - conda-forge
-  - nodefaults
-dependencies:
-  - cmap = 0.6.2
-  - country_converter = 1.3.2
-  - geopandas = 1.0.1
-  - matplotlib = 3.10.8
-  - numpy = 2.3.4
-  - pandas = 2.3.3
-  - pandera-geopandas = 0.24
-  - pandera-io = 0.24
-  - pyarrow = 19.0.1
-  - python = 3.13.9
-  - scipy = 1.16.3
-  - shapely = 2.1.2
