+++ workflow/rules/automatic.smk
@@ -25,7 +25,7 @@ rule download_sci_grid:
         "<logs>/automatic/download_sci_grid.log",
     localrule: True
     conda:
-        "../envs/shell.yaml"
+        "../envs/module.yaml"
     params:
         url=internal["resources"]["automatic"]["scigrid_gas"],
     message:
@@ -42,7 +42,7 @@ rule unzip_scigrid_dataset:
     log:
         "<logs>/automatic/unzip_scigrid_dataset_{scigrid_gas}.log",
     conda:
-        "../envs/euro-gas-grid.yaml"
+        "../envs/module.yaml"
     params:
         file=lambda wc: f"data/IGGIELGNC3_{wc.scigrid_gas}.geojson",
     message:
@@ -58,7 +58,7 @@ rule download_salt_cavern_storage:
         "<logs>/automatic/download_salt_cavern_storage.log",
     localrule: True
     conda:
-        "../envs/shell.yaml"
+        "../envs/module.yaml"
     params:
         url=internal["resources"]["automatic"]["salt_cavern_h2"],
     message:
@@ -74,7 +74,7 @@ rule download_natural_earth:
         "<logs>/automatic/download_{nat_earth}.log",
     localrule: True
     conda:
-        "../envs/shell.yaml"
+        "../envs/module.yaml"
     params:
         url=lambda wc: internal["resources"]["automatic"]["natural_earth"][wc.nat_earth],
     message:
